Alpha as of 30/12/2025 - Implemented W.I.P customised horizontal scroll bar - comprised of 4 controls, being: - ScrollBarBgPnl - the background panel of the scroll bar - LeftScrollBtn & RightScrollBtn - the buttons that shift the scrollbar - ScrollBarPnl - the scroll bar itself.
This new horizontal scroll bar currently is not finished - it currently only appears. The buttons throw a NotImplementedException, and the scroll bar has no functionality. This was added to be able to add images to each part of the scroll bar. However, the custom scroll bar doesn't work like the default windows one. The scroll bar stays the same size, so the distance from the left of the scroll bar is scaled to create the actual offset in pixels. Example: If the needed change from the scrollbar min-max is 0, the scrollbar will still be the same size and you will still be able to scroll it, but when it is scaled the offset will always be 0 (since the needed change is 0). If the min-max was a different value, then the same principles are still in place.

Other changes include:
- New notebook button in Digiffice Allnote file ribbon tab has a background image.
- a second button in the Digiffice Allnote file ribbon tab. Also has image.
- The Allnote UI has been shuffled around. Instead of a page in the middle with the sections, list of pages etc. being on the left of the screen, the list of pages are now on the right (in the spare space in sectionBG) and the page itself is in sectionBG. The left of the screen will now include only sections and section groups.